backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.1
certifi==2026.2.27
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.12
